requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
python-multipart==0.0.6

# WebSocket support
websockets==12.0

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Process management
psutil==5.9.6
watchdog==3.0.0

# RAG system dependencies
redis==5.0.1
ratelimit==2.2.1

# Google Cloud / Vertex AI (optional, for knowledge base)
google-cloud-aiplatform
google-cloud-storage
vertexai

# Vector store dependencies (optional)
faiss-cpu
numpy
pinecone-client==2.2.4

# NLP and text processing
beautifulsoup4==4.12.2
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
python-json-logger==2.0.7
prometheus-client==0.19.0